Menambahkan Api Key ke Semua Method

+++ ECommerceAPI/src/OrdersHandler.java
@@ -1,5 +1,6 @@
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
+import io.github.cdimascio.dotenv.Dotenv;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -10,6 +11,13 @@ import java.util.Map;
 import java.util.HashMap;
 
 public class OrdersHandler implements HttpHandler {
+    private static final String API_KEY;
+
+    static {
+        Dotenv dotenv = Dotenv.configure().directory(".env").load();
+        API_KEY = dotenv.get("API_KEY");
+    }
+
     private Connection getConnection() throws SQLException {
         return DatabaseConnection.connect();
     }
@@ -53,6 +61,11 @@ public class OrdersHandler implements HttpHandler {
         return params;
     }
 
+    private static boolean validateApiKey(HttpExchange exchange) {
+        String apiKey = exchange.getRequestHeaders().getFirst("x-api-key");
+        return apiKey != null && apiKey.equals(API_KEY);
+    }
+
 
     @Override
     public void handle(HttpExchange exchange) throws IOException {
@@ -82,10 +95,10 @@ public class OrdersHandler implements HttpHandler {
     }
 
     private void handleGetAllOrders(HttpExchange exchange) throws IOException {
-        //            if (!validateApiKey(exchange)) {
-//                sendErrorResponse(exchange, 401, "Unauthorized");
-//                return;
-//            }
+        if (!validateApiKey(exchange)) {
+                sendErrorResponse(exchange, 401, "Unauthorized");
+                return;
+            }
 
         // Mendapatkan nilai query params "type" dari URL
         String query = exchange.getRequestURI().getQuery();
@@ -152,6 +165,10 @@ public class OrdersHandler implements HttpHandler {
     }
 
     private void handleGetOrder(HttpExchange exchange) throws IOException {
+        if (!validateApiKey(exchange)) {
+            sendErrorResponse(exchange, 401, "Unauthorized");
+            return;
+        }
         String orderIdString = extractOrderId(exchange.getRequestURI().getPath());
         if (orderIdString != null) {
             try {
@@ -218,6 +235,10 @@ public class OrdersHandler implements HttpHandler {
     }
 
     private void handleCreateOrder(HttpExchange exchange) throws IOException {
+        if (!validateApiKey(exchange)) {
+            sendErrorResponse(exchange, 401, "Unauthorized");
+            return;
+        }
         String requestBody = getRequestData(exchange);
         try {
             JSONObject orderObject = new JSONObject(requestBody);
@@ -311,6 +332,10 @@ public class OrdersHandler implements HttpHandler {
     }
 
     private void handleDeleteOrder(HttpExchange exchange, int productId) throws IOException {
+        if (!validateApiKey(exchange)) {
+            sendErrorResponse(exchange, 401, "Unauthorized");
+            return;
+        }
         try (Connection connection = getConnection()){
             try (PreparedStatement statement = connection.prepareStatement(
                     "DELETE FROM orders WHERE order_id = ?")) {
